requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
whitenoise==6.5.0
drf-spectacular==0.28.0
schemas==0.7.1
pytest==8.3.5
pytest-django==4.10.0